Changed the signature of Arr::{typeCheckLooseStructure, TypeCheckStrictStructure} methods
The change refactors the `$additionalArgsType: Option(?Type)` argument into `$extraKeysType: ?Type`.  The null option is included to allow a default value of `null`, which is immediately converted to Types::nothing(), which causes no additional keys to be accepted (since no value type-checks as an instance of `nothing`).

This change simplifies the signatures and calls to the method, and allows the type constructor to use only types to define its behavior, rather than including an Option structure wrapping it when none is required.